osint-tool/requirements.txt
# Python dependencies
requests
whois
dns-python
pillow
tabulate
